LPS-50418 Move viewAction into Portlet to be executed before render
