+++ modules/apps/nested-portlets/nested-portlets-web/src/com/liferay/nested/portlets/web/portlet/NestedPortletsPortlet.java
@@ -14,17 +14,47 @@
 
 package com.liferay.nested.portlets.web.portlet;
 
+import com.liferay.nested.portlets.web.portlet.util.NestedPortletUtil;
 import com.liferay.nested.portlets.web.upgrade.NestedPortletUpgrade;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCPortlet;
+import com.liferay.portal.kernel.util.ArrayUtil;
+import com.liferay.portal.kernel.util.StringBundler;
+import com.liferay.portal.kernel.util.StringPool;
+import com.liferay.portal.kernel.util.StringUtil;
+import com.liferay.portal.kernel.util.UnicodeProperties;
+import com.liferay.portal.kernel.util.Validator;
+import com.liferay.portal.model.Layout;
+import com.liferay.portal.model.LayoutTemplate;
+import com.liferay.portal.model.LayoutTemplateConstants;
+import com.liferay.portal.model.LayoutTypePortletConstants;
+import com.liferay.portal.model.Theme;
+import com.liferay.portal.service.LayoutLocalServiceUtil;
+import com.liferay.portal.service.LayoutTemplateLocalServiceUtil;
+import com.liferay.portal.theme.ThemeDisplay;
+import com.liferay.portal.util.WebKeys;
+
+import java.io.IOException;
+
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+import javax.portlet.Portlet;
+import javax.portlet.PortletException;
+import javax.portlet.PortletPreferences;
+import javax.portlet.RenderRequest;
+import javax.portlet.RenderResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
-import javax.portlet.Portlet;
-
 /**
  * @author Peter Fellwock
-*/
+ */
 @Component(
 	immediate = true,
 	property = {
@@ -48,12 +78,146 @@ import javax.portlet.Portlet;
 		"javax.portlet.resource-bundle=content.Language",
 		"javax.portlet.security-role-ref=guest,power-user,user"
 	},
-	service = Portlet.class	
+	service = Portlet.class
 )
 public class NestedPortletsPortlet extends MVCPortlet {
 
-	@Reference(unbind = "-")
-	protected void setNestedPortletUpgrade(NestedPortletUpgrade nestedPortletUpgrade) {
+	@Override
+	public void doView(
+			RenderRequest renderRequest, RenderResponse renderResponse)
+		throws IOException, PortletException {
+
+		ThemeDisplay themeDisplay = (ThemeDisplay)renderRequest.getAttribute(
+			WebKeys.THEME_DISPLAY);
+
+		PortletPreferences portletPreferences = renderRequest.getPreferences();
+
+		String layoutTemplateId = NestedPortletUtil.getLayoutTemplateId(
+			portletPreferences);
+
+		String velocityTemplateId = StringPool.BLANK;
+		String velocityTemplateContent = StringPool.BLANK;
+
+		Map<String, String> columnIds = new HashMap<String, String>();
+
+		if (Validator.isNotNull(layoutTemplateId)) {
+			Theme theme = themeDisplay.getTheme();
+
+			LayoutTemplate layoutTemplate =
+				LayoutTemplateLocalServiceUtil.getLayoutTemplate(
+					layoutTemplateId, false, theme.getThemeId());
+
+			String content = layoutTemplate.getContent();
+
+			Matcher processColumnMatcher = _processColumnPattern.matcher(
+				content);
+
+			while (processColumnMatcher.find()) {
+				String columnId = processColumnMatcher.group(2);
+
+				if (Validator.isNotNull(columnId)) {
+					columnIds.put(
+						columnId, renderResponse.getNamespace() +
+						StringPool.UNDERLINE + columnId);
+				}
+			}
+
+			processColumnMatcher.reset();
+
+			StringBundler sb = new StringBundler(4);
+
+			sb.append(theme.getThemeId());
+			sb.append(LayoutTemplateConstants.CUSTOM_SEPARATOR);
+			sb.append(renderResponse.getNamespace());
+			sb.append(layoutTemplateId);
+
+			velocityTemplateId = sb.toString();
+
+			content = processColumnMatcher.replaceAll("$1\\${$2}$3");
+
+			Matcher columnIdMatcher = _columnIdPattern.matcher(content);
+
+			velocityTemplateContent =
+				columnIdMatcher.replaceAll(
+					"$1" +renderResponse.getNamespace() + "$2$3");
+		}
+
+		checkLayout(themeDisplay.getLayout(), columnIds.values());
+
+		renderRequest.setAttribute(
+			WebKeys.NESTED_PORTLET_VELOCITY_TEMPLATE_ID, velocityTemplateId);
+		renderRequest.setAttribute(
+				WebKeys.NESTED_PORTLET_VELOCITY_TEMPLATE_CONTENT,
+				velocityTemplateContent);
+
+		@SuppressWarnings("unchecked")
+		Map<String, Object> vmVariables =
+			(Map<String, Object>)renderRequest.getAttribute(
+				WebKeys.VM_VARIABLES);
+
+		if (vmVariables != null) {
+			vmVariables.putAll(columnIds);
+		}
+		else {
+			renderRequest.setAttribute(WebKeys.VM_VARIABLES, columnIds);
+		}
+
+		super.include(viewTemplate, renderRequest, renderResponse);
 	}
+
+	protected void checkLayout(Layout layout, Collection<String> columnIds) {
+		UnicodeProperties properties = layout.getTypeSettingsProperties();
+
+		String[] layoutColumnIds =
+			StringUtil.split(
+				properties.get(LayoutTypePortletConstants.NESTED_COLUMN_IDS));
+
+		boolean updateColumnIds = false;
+
+		for (String columnId : columnIds) {
+			String portletIds = properties.getProperty(columnId);
+
+			if (Validator.isNotNull(portletIds) &&
+				!ArrayUtil.contains(layoutColumnIds, columnId)) {
+
+				layoutColumnIds = ArrayUtil.append(layoutColumnIds, columnId);
+
+				updateColumnIds = true;
+			}
+		}
+
+		if (updateColumnIds) {
+			properties.setProperty(
+					LayoutTypePortletConstants.NESTED_COLUMN_IDS,
+					StringUtil.merge(layoutColumnIds));
+
+			layout.setTypeSettingsProperties(properties);
+
+			try {
+				LayoutLocalServiceUtil.updateLayout(
+					layout.getGroupId(), layout.isPrivateLayout(),
+					layout.getLayoutId(), layout.getTypeSettings());
+			}
+			catch (Exception e) {
+				if (_log.isWarnEnabled()) {
+					_log.warn(e, e);
+				}
+			}
+		}
+	}
+
+	@Reference(unbind = "-")
+	protected void setNestedPortletUpgrade(
+		NestedPortletUpgrade nestedPortletUpgrade) {
+	}
+
+	private static Log _log = LogFactoryUtil.getLog(
+		NestedPortletsPortlet.class);
+
+	private static Pattern _columnIdPattern = Pattern.compile(
+		"([<].*?id=[\"'])([^ ]*?)([\"'].*?[>])", Pattern.DOTALL);
+	private static Pattern _processColumnPattern = Pattern.compile(
+			"(processColumn[(]\")(.*?)(\"(?:, *\"(?:.*?)\")?[)])",
+			Pattern.DOTALL);
 
 }
